converter/requirements.txt
fastapi
pydantic
requests
python-docx
unidecode
uvicorn
python-multipart